fix(printers): address Copilot review feedback on PR #105
Ten inline comments, all addressed:

- PrintingReportsController: clamp year/month from request input so
  Carbon::create doesn't 500 on invalid values; eager-load the
  polymorphic assignee's department via morphWith(User::class) to
  drop an N+1 on fleets with many printers.
- /reports/printing: link asset name to `#printing` anchor so users
  land directly on the per-asset Printing tab; render cost cells via
  Helper::formatCurrencyOutput for currency-config consistency.
- Per-asset Printing tab: same Helper::formatCurrencyOutput swap on
  the last-30-days cost card, single-GL fallback, and recent-jobs cost
  column.
- PrinterUsageService::fleetMonth (MySQL path): align refund detection
  with the PHP looksLikeRefund() heuristic -- includes the 'y' truthy
  value and the "transaction type LIKE '%REFUND%'" fallback, so fleet
  rates match per-asset rates.
- PrinterUsageService::monthlyVolume: capture Carbon::now() once and
  derive every month from that anchor; the loop previously called
  Carbon::now() on each iteration, risking inconsistency on a
  month-boundary clock tick during a request.
- Migration: switch printer_asset_id from unsignedBigInteger to
  unsignedInteger to match the actual assets.id column type. Smaller
  index, and a real FK constraint becomes possible later.

+++ database/migrations/2026_05_26_000001_add_printer_asset_id_to_transaction_raw_rows.php
@@ -9,7 +9,10 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('transaction_raw_rows', function (Blueprint $table) {
-            $table->unsignedBigInteger('printer_asset_id')->nullable()->after('source_kind');
+            // `assets.id` is unsignedInteger, not unsignedBigInteger -- keep
+            // the FK column the same width so the index is smaller and a real
+            // foreign-key constraint can be added later if needed.
+            $table->unsignedInteger('printer_asset_id')->nullable()->after('source_kind');
             $table->index(
                 ['printer_asset_id', 'period_year', 'period_month'],
                 'transaction_raw_rows_printer_period'
